+++ Assets/VRM10_Samples/VRM10Viewer/VRM10Loaded.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0565f187ccc74bf41a4ce6d534422efa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
